+++ lib/rails_stuff/resources_controller/actions.rb
@@ -8,21 +8,21 @@ module RailsStuff
 
       def create(options = {}, &block)
         if create_resource
-          options[:location] = after_save_url
+          options[:location] ||= after_save_url
         end
         respond_with(resource, options, &block)
       end
 
       def update(options = {}, &block)
         if update_resource
-          options[:location] = after_save_url
+          options[:location] ||= after_save_url
         end
         respond_with(resource, options, &block)
       end
 
       def destroy(options = {}, &block)
         resource.destroy
-        options[:location] = after_destroy_url
+        options[:location] ||= after_destroy_url
         flash_errors!
         respond_with(resource, options, &block)
       end